int64 -> int

+++ api_gateway/alice-router/handlers/middleware.go
@@ -9,7 +9,7 @@ import (
 type AuthResponse struct {
 	AccessToken  string `json:"access_token"`
 	RefreshToken string `json:"refresh_token"`
-	UserID       int64  `json:"user_id"`
+	UserID       int    `json:"user_id"`
 }
 
 type LoginRequest struct {
